requirements.txt
numpy
scipy
psycopg2-binary
python-dotenv
nltk
xgboost
streamlit
pyyaml
joblib
scikit-learn
apache-airflow-providers-postgres